+++ src/LevelDMD.cpp
@@ -11,7 +11,7 @@ LevelDMD::LevelDMD(uint16_t width, uint16_t height, bool sam)
 {
   m_width = width;
   m_height = height;
-  m_length = width * height;
+  m_length = (int)width * height;
   m_pitch = width;
   m_sam = sam;
